Embed CampaignTrendRow; refactor service & DAO
Move CampaignTrendRow into CampaignTrendResponse (remove separate file) and update imports/usages.
Simplify CampaignReportService by inlining reactive DAO calls for campaign IDs and metrics (uses campaignDAO.findCampaignIdsForProducts and campaignMetricDAO.findByFilters) and remove several helper methods.
Streamline ProductDAO.getAllProducts by inlining condition handling and using processorAccessCondition with FlatMapUtil.
Overall cleanup consolidates model classes and simplifies reactive data retrieval paths.
